chore: remove demo data and add fallbacks
