--- .env.local.example
@@ -1,19 +1,0 @@
-# Defaults, used by ./intro-template and can be deleted if the component is removed
-NEXT_PUBLIC_VERCEL_GIT_REPO_OWNER="sanity-io"
-NEXT_PUBLIC_VERCEL_GIT_PROVIDER="github"
-NEXT_PUBLIC_VERCEL_GIT_REPO_SLUG="nextjs-blog-cms-sanity-v3"
-
-# Required, find them on https://manage.sanity.io
-NEXT_PUBLIC_SANITY_PROJECT_ID=
-NEXT_PUBLIC_SANITY_DATASET=
-# see https://www.sanity.io/docs/api-versioning for how versioning works
-NEXT_PUBLIC_SANITY_API_VERSION="2022-11-15"
-SANITY_API_READ_TOKEN=
-# Optional, useful if you plan to add API functions that can write to your dataset
-SANITY_API_WRITE_TOKEN=
-
-# Optional, can be used to change the Studio title in the navbar and differentiate between production and staging environments for your editors
-NEXT_PUBLIC_SANITY_PROJECT_TITLE="Next.js Blog with Sanity.io"
-
-# Optional, check the comments in pages/api/revalidate.ts for instructions on how to set it up
-SANITY_REVALIDATE_SECRET=
