+++ components/PostPageHead.tsx
@@ -1,5 +1,5 @@
 import BlogMeta from 'components/BlogMeta'
-import * as demo from 'lib/demo.data'
+import * as fallback from 'lib/fallback.data'
 import { urlForImage } from 'lib/sanity.image'
 import { Post, Settings } from 'lib/sanity.queries'
 import Head from 'next/head'
@@ -10,7 +10,7 @@ export interface PostPageHeadProps {
 }
 
 export default function PostPageHead({ settings, post }: PostPageHeadProps) {
-  const title = settings.title ?? demo.title
+  const title = settings.title ?? fallback.title
   return (
     <Head>
       <title>{post.title ? `${post.title} | ${title}` : title}</title>
