+++ lib/fallback.data.ts
@@ -1,0 +1,23 @@
+// All the demo data that used as fallbacks when there's nothing in the dataset yet
+
+export const title = 'Sammy Dowds.'
+
+export const description = [
+  {
+    _key: '9f1a629887fd',
+    _type: 'block',
+    children: [
+      {
+        _key: '4a58edd077880',
+        _type: 'span',
+        marks: [],
+        text: 'Building, learning, and sharing.',
+      },
+    ],
+    markDefs: [
+    ],
+    style: 'normal',
+  },
+]
+
+export const ogImageTitle = "Sammy Dowds's Engineering Blog"
