+++ pages/api/og.tsx
@@ -7,7 +7,7 @@ import { createClient } from 'next-sanity'
 export const config: PageConfig = { runtime: 'edge' }
 
 import { height, OpenGraphImage, width } from 'components/OpenGraphImage'
-import * as demo from 'lib/demo.data'
+import * as fallback from 'lib/fallback.data'
 import { Settings, settingsQuery } from 'lib/sanity.queries'
 
 export default async function og(req: NextRequest, res: NextResponse) {
@@ -29,7 +29,7 @@ export default async function og(req: NextRequest, res: NextResponse) {
   }
 
   return new ImageResponse(
-    <OpenGraphImage title={title || demo.ogImageTitle} />,
+    <OpenGraphImage title={title || fallback.ogImageTitle} />,
     {
       width,
       height,
