+++ schemas/settings/index.ts
@@ -1,5 +1,5 @@
 import { CogIcon } from '@sanity/icons'
-import * as demo from 'lib/demo.data'
+import * as fallback from 'lib/fallback.data'
 import { defineArrayMember, defineField, defineType } from 'sanity'
 
 import OpenGraphInput from './OpenGraphInput'
@@ -18,7 +18,7 @@ export default defineType({
       description: 'This field is the title of your blog.',
       title: 'Title',
       type: 'string',
-      initialValue: demo.title,
+      initialValue: fallback.title,
       validation: (rule) => rule.required(),
     }),
     defineField({
@@ -27,7 +27,7 @@ export default defineType({
         'Used both for the <meta> description tag for SEO, and the blog subheader.',
       title: 'Description',
       type: 'array',
-      initialValue: demo.description,
+      initialValue: fallback.description,
       of: [
         defineArrayMember({
           type: 'block',
@@ -69,7 +69,7 @@ export default defineType({
           name: 'title',
           title: 'Title',
           type: 'string',
-          initialValue: demo.ogImageTitle,
+          initialValue: fallback.ogImageTitle,
         }),
       ],
     }),
